PT-2289 - Allow pt-stalk do disable ps-lock-transactions data collection via parameter
- Implemented feature
- Started working on the test

+++ lib/bash/collect.sh
@@ -243,8 +243,10 @@ collect_mysql_data_one() {
    # get and keep a connection to the database; in troubled times
    # the database tends to exceed max_connections, so reconnecting
    # in the loop tends not to work very well.
-   $CMD_MYSQLADMIN $EXT_ARGV ext -i$OPT_SLEEP_COLLECT -c$cnt >>"$d/$p-mysqladmin" &
-   mysqladmin_pid=$!
+   if ! [ "${OPT_SKIP_COLLECTION}" =~ "mysqladmin" ]; then
+      $CMD_MYSQLADMIN $EXT_ARGV ext -i$OPT_SLEEP_COLLECT -c$cnt >>"$d/$p-mysqladmin" &
+      mysqladmin_pid=$!
+   fi
 
    ps_instrumentation_enabled=$($CMD_MYSQL $EXT_ARGV -e 'SELECT ENABLED FROM performance_schema.setup_instruments WHERE NAME = "transaction";' \
                                       | sed "2q;d" | sed 'y/ABCDEFGHIJKLMNOPQRSTUVWXYZ/abcdefghijklmnopqrstuvwxyz/')
@@ -289,18 +291,25 @@ collect_mysql_data_loop() {
 
    # SHOW FULL PROCESSLIST duplicates information in performance_schema.threads we collecting now
    # Keeping it for backward compatibility and may remove in the future
-
+   if ! [ "${OPT_SKIP_COLLECTION}" =~ "processlist" ]; then
       (echo $ts; $CMD_MYSQL $EXT_ARGV -e "SHOW FULL PROCESSLIST\G") \
       >> "$d/$p-processlist" &
-      (echo $ts; $CMD_MYSQL $EXT_ARGV -e "SELECT * FROM performance_schema.threads\G") \
-      >> "$d/$p-threads" &
-
-   if [ "$have_lock_waits_table" ]; then
-      (echo $ts; lock_waits "$d/lock_waits.running")   >>"$d/$p-lock-waits" &
-      (echo $ts; transactions) >>"$d/$p-transactions" &
    fi
 
-   if [ "${mysql_version}" '>' "5.6" ] && [ $ps_instrumentation_enabled == "yes" ]; then
+   (echo $ts; $CMD_MYSQL $EXT_ARGV -e "SELECT * FROM performance_schema.threads\G") \
+   >> "$d/$p-threads" &
+
+   if [ "$have_lock_waits_table" ]; then
+      if ! [ "${OPT_SKIP_COLLECTION}" =~ "lock-waits" ]; then
+         (echo $ts; lock_waits "$d/lock_waits.running")   >>"$d/$p-lock-waits" &
+      fi
+      if ! [ "${OPT_SKIP_COLLECTION}" =~ "transactions" ]; then
+         (echo $ts; transactions) >>"$d/$p-transactions" &
+      fi
+   fi
+
+   if [ "${mysql_version}" '>' "5.6" ] && [ $ps_instrumentation_enabled == "yes" ] \
+      && ! [ "${OPT_SKIP_COLLECTION}" =~ "ps-locks-transactions" ]; then
       ps_locks_transactions "$d/$p-ps-locks-transactions"
    fi
 
@@ -538,20 +547,22 @@ innodb_status() {
 
    local innostat=""
 
-   $CMD_MYSQL $EXT_ARGV -e "SHOW /*!40100 ENGINE*/ INNODB STATUS\G" \
-      >> "$d/$p-innodbstatus$n"
-   grep "END OF INNODB" "$d/$p-innodbstatus$n" >/dev/null || {
-      if [ -d /proc -a -d /proc/$mysqld_pid ]; then
-         for fd in /proc/$mysqld_pid/fd/*; do
-            file $fd | grep deleted >/dev/null && {
-               grep 'INNODB' $fd >/dev/null && {
-                  cat $fd > "$d/$p-innodbstatus$n"
-                  break
+   if ! [ "${OPT_SKIP_COLLECTION}" =~ "innodbstatus" ]; then
+      $CMD_MYSQL $EXT_ARGV -e "SHOW /*!40100 ENGINE*/ INNODB STATUS\G" \
+         >> "$d/$p-innodbstatus$n"
+      grep "END OF INNODB" "$d/$p-innodbstatus$n" >/dev/null || {
+         if [ -d /proc -a -d /proc/$mysqld_pid ]; then
+            for fd in /proc/$mysqld_pid/fd/*; do
+               file $fd | grep deleted >/dev/null && {
+                  grep 'INNODB' $fd >/dev/null && {
+                     cat $fd > "$d/$p-innodbstatus$n"
+                     break
+                  }
                }
-            }
-         done
-      fi
-   }
+            done
+         fi
+      }
+   fi
 }
 
 rocksdb_status() {
@@ -560,7 +571,7 @@ rocksdb_status() {
     has_rocksdb=`$CMD_MYSQL $EXT_ARGV -e "SHOW ENGINES" | grep -i 'rocksdb'`
     exit_code=$?
 
-    if [ $exit_code -eq 0 ]; then
+    if [ $exit_code -eq 0 ] && ! [ "${OPT_SKIP_COLLECTION}" =~ "rocksdbstatus" ]; then
         $CMD_MYSQL $EXT_ARGV -e "SHOW ENGINE ROCKSDB STATUS\G" \
                    >> "$d/$p-rocksdbstatus$n" || rm -f "$d/$p-rocksdbstatus$n"
     fi
@@ -646,17 +657,19 @@ collect_mysql_variables() {
    echo -e "\n$sql\n" >> $outfile
    $CMD_MYSQL $EXT_ARGV -e "$sql" >> $outfile
 
-   sql="select * from performance_schema.variables_by_thread order by thread_id, variable_name;"
-   echo -e "\n$sql\n" >> $outfile
-   $CMD_MYSQL $EXT_ARGV -e "$sql" >> $outfile
+   if ! [ "${OPT_SKIP_COLLECTION}" =~ "thread-variables" ]; then
+      sql="select * from performance_schema.variables_by_thread order by thread_id, variable_name;"
+      echo -e "\n$sql\n" >> $outfile
+      $CMD_MYSQL $EXT_ARGV -e "$sql" >> $outfile
 
-   sql="select * from performance_schema.user_variables_by_thread order by thread_id, variable_name;"
-   echo -e "\n$sql\n" >> $outfile
-   $CMD_MYSQL $EXT_ARGV -e "$sql" >> $outfile
+      sql="select * from performance_schema.user_variables_by_thread order by thread_id, variable_name;"
+      echo -e "\n$sql\n" >> $outfile
+      $CMD_MYSQL $EXT_ARGV -e "$sql" >> $outfile
 
-   sql="select * from performance_schema.status_by_thread order by thread_id, variable_name; "
-   echo -e "\n$sql\n" >> $outfile
-   $CMD_MYSQL $EXT_ARGV -e "$sql" >> $outfile
+      sql="select * from performance_schema.status_by_thread order by thread_id, variable_name; "
+      echo -e "\n$sql\n" >> $outfile
+      $CMD_MYSQL $EXT_ARGV -e "$sql" >> $outfile
+   fi
 
 }
